move osmnx readers into separate file
